Avoiding rounding errors, made bone-cache-clear a separate function to render

+++ src/main/java/moe/nightfall/instrumentality/PoseBoneTransform.java
@@ -25,14 +25,14 @@ public class PoseBoneTransform {
      * The reason for having 3 sets of values is to allow some rotations to happen after/before others.
      * (for example: rotation on one axis to raise/lower head should be applied after left/right rotation)
      */
-    public float X0, Y0, Z0;
-    public float X1, Y1;
-    public float X2;
+    public double X0, Y0, Z0;
+    public double X1, Y1;
+    public double X2;
 
     /**
      * The translation values(applied before rotation)
      */
-    public float TX0, TY0, TZ0;
+    public double TX0, TY0, TZ0;
 
     public PoseBoneTransform() {
 
@@ -86,13 +86,13 @@ public class PoseBoneTransform {
     }
 
     public void apply(Matrix4f boneMatrix) {
-        boneMatrix.translate(new Vector3f(TX0, TY0, TZ0));
+        boneMatrix.translate(new Vector3f((float) TX0, (float) TY0, (float) TZ0));
 
-        boneMatrix.rotate(X0, new Vector3f(1, 0, 0));
-        boneMatrix.rotate(Y0, new Vector3f(0, 1, 0));
-        boneMatrix.rotate(Z0, new Vector3f(0, 0, 1));
-        boneMatrix.rotate(X1, new Vector3f(1, 0, 0));
-        boneMatrix.rotate(Y1, new Vector3f(0, 1, 0));
-        boneMatrix.rotate(X2, new Vector3f(1, 0, 0));
+        boneMatrix.rotate((float) X0, new Vector3f(1, 0, 0));
+        boneMatrix.rotate((float) Y0, new Vector3f(0, 1, 0));
+        boneMatrix.rotate((float) Z0, new Vector3f(0, 0, 1));
+        boneMatrix.rotate((float) X1, new Vector3f(1, 0, 0));
+        boneMatrix.rotate((float) Y1, new Vector3f(0, 1, 0));
+        boneMatrix.rotate((float) X2, new Vector3f(1, 0, 0));
     }
 }
